LPS-116972 Moves to zip package
