+++ modules/apps/style-book/style-book-web/src/main/java/com/liferay/style/book/web/internal/portlet/action/ExportStyleBookEntriesMVCResourceCommand.java
@@ -23,7 +23,7 @@ import com.liferay.portal.kernel.util.Time;
 import com.liferay.style.book.constants.StyleBookPortletKeys;
 import com.liferay.style.book.model.StyleBookEntry;
 import com.liferay.style.book.service.StyleBookEntryLocalService;
-import com.liferay.style.book.web.internal.portlet.helper.ExportHelper;
+import com.liferay.style.book.web.internal.portlet.zip.ExportHelper;
 
 import java.io.File;
 import java.io.FileInputStream;
