+++ modules/apps/style-book/style-book-web/src/main/resources/META-INF/resources/init.jsp
@@ -36,7 +36,7 @@ page import="com.liferay.style.book.web.internal.display.context.EditStyleBookEn
 page import="com.liferay.style.book.web.internal.display.context.ImportStyleBookDisplayContext" %><%@
 page import="com.liferay.style.book.web.internal.display.context.StyleBookDisplayContext" %><%@
 page import="com.liferay.style.book.web.internal.display.context.StyleBookManagementToolbarDisplayContext" %><%@
-page import="com.liferay.style.book.web.internal.portlet.helper.StyleBookImporterResultEntry" %><%@
+page import="com.liferay.style.book.web.internal.portlet.zip.StyleBookImporterResultEntry" %><%@
 page import="com.liferay.style.book.web.internal.servlet.taglib.clay.StyleBookVerticalCard" %>
 
 <%@ page import="java.util.List" %>
